[IMP] point_of_sale: use more unique identifier
When custom modules load a different partner field set for a completely
unrelated purpose AND this loading takes place prior to the stock odoo
one, then the load_new_partners method will infer the wrong field set
as the first in row hit is returned by _.find.

It is not optimal code quality to use `res.partner` as if it was a
unique identifier, while not inhibiting multiple loads of on a single
model.

Adding a (more) unique label and refer it in the find solves this
problem and makes the code more robust.

closes odoo/odoo#32351

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>