[FIX] website: allow empty strings in title and meta description
Before this commit, In SEO dialog once you set the title and description,
it was not possible to reset title and description. The reason behind the
issue was, update method was taking title and description from DOM
instead of dialog inputs.

After this commit, you will be able to save empty strings in the title and
description. Once user will do this, the page will use default title.

Task ID:1949636